+++ src/entities_game_service/player_manager/cdb.h
@@ -137,7 +137,7 @@ public:
 		/**
 		 *	Return the count of strings composing this id
 		 */
-		uint size() const { return _Ids.size(); }
+		uint size() const { return (uint)_Ids.size(); }
 
 	private:
 		std::vector<std::string> _Ids;
@@ -327,7 +327,7 @@ protected:
 	/// Atomic flag: is the branch an atomic group, or is the leaf a member of an atomic group
 	bool					_Atomic;
 
-	/// Optionnal callback pointer.
+	/// Optional callback pointer.
 	TNodeChangeCallback		_ChangeCallback;
 
 };
